AB-2 feat(jira): Refactor Jira polling logic into a new module supporting multiple projects
